Restructure user settings with integrated faction management
- Consolidate faction management into main user settings page
- Replace separate /private/user/factions page with inline functionality
- Add comprehensive faction CRUD operations within factions tab
- Implement proper form state management and data invalidation
- Add faction management validation schema to lib/validation.ts
- Integrate add/edit/delete faction forms with real-time list updates
- Remove obsolete standalone factions page and server files
- Improve UX with single-page settings experience and better mobile support
